refactor: extract order detail cards and unify transactions cart hook
